requirements.txt
Flask==2.0.1
pandas==1.3.3
numpy==1.21.2
scikit-learn==0.24.2
matplotlib==3.4.3
seaborn==0.11.2
gunicorn==20.1.0
haversine
Flask-SocketIO==5.1.0
Flask-CORS==3.0.10
reportlab==3.5.67
openpyxl==3.0.7
Flask==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Cors==3.0.10
Flask-SocketIO==5.1.0
Werkzeug==2.0.3
PyJWT==2.3.0
pandas==1.3.4
numpy==1.21.2
scikit-learn==1.0.1
osmnx==1.1.1
fPDF==1.7.2
python-decouple==3.5
